backend/requirements.txt
google-generativeai
python-dotenv
feedparser
newspaper3k
langchain
streamlit
faiss-cpu
numpy
yfinance
pandas_ta
langchain-community
langchain-openai
langchain-google-genai
fastapi
pydantic
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2